created layout folder, and moved total dosage detail to own component
